abstract Form, polish
